Add files via upload
This code is for greeting the user, asking them what they need assistance with and then completing the task based on their response.  This script is still in development and only basic use such as time, weather and news.   I am still need to consult how to find the headline news for when this task gets asked.